api/backend: backend_command doesn't need to be here
It's not part of the public backend API.
